Fixed my classes

+++ src/main/java/za/ac/cput/Repository/VehicleDiscRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import za.ac.cput.Domain.bookings.VehicleDisc;
 @Repository
-public interface VehicleDiscRepository extends JpaRepository<VehicleDisc, Integer> {
+public interface VehicleDiscRepository extends JpaRepository<VehicleDisc, Long> {
 }
