+++ src/main/java/za/ac/cput/Service/IVehicleDiscService.java
@@ -4,6 +4,6 @@ import za.ac.cput.Domain.bookings.VehicleDisc;
 
 import java.util.List;
 
-public interface IVehicleDiscService extends IService <VehicleDisc, Integer> {
+public interface IVehicleDiscService extends IService <VehicleDisc, Long> {
     List<VehicleDisc> getAll();
 }
